update support for new repos
in some cases, the update command needs to install new repos
from scratch. This is to support the case where you need to
use librarian-puppet to update a cache by adding a new repo.

this code ensures that update will run install in the case
where the target directory does not already exist.