Change the fetch procedure of the installer
The installer uses checkout to switch to the specified branch.
This method usually works, but is not compatible with Gerrit's
refs/changes/* branches. Using "git fetch origin $ref" and
then "git checkout FETCH_HEAD" works with both normal and
Gerrit branches and is already being used for the "update"
action.